Remove usage of del function, refactor tests
